modificaciones para asignaciones de tareas y proyectos

+++ src/main/java/com/uade/consultancymanager/repository/ProyectoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProyectoRepository extends JpaRepository<Proyectos, Long> {
+public interface ProyectoRepository extends JpaRepository<Proyectos, Integer> {
     boolean existsByNombre(String nombre);
 }
